refactor(db): Extract collection operations into collection_ops.rs
Phase 3 of the database.rs refactoring plan. Extract all collection
CRUD and query operations into a dedicated module.

Extracted operations:
- load_collections: Load all collections from storage
- reload_cache: Refresh collection cache from storage
- next_collection_short_id: Sequence ID generation
- create_collection_with_txn/create_collection/create_collections_atomic
- delete_collection_with_txn/delete_collection
- truncate_collection/truncate_collection_inner
- list_collections_with_txn/list_collections
- get_collection_with_txn/get_collection
- has_collection_with_txn/has_collection
- find_collection_by_id
- collections_snapshot
- set_active_collection_version
- get_collection_by_version_id/get_collection_by_version_id_full
- get_all_collection_versions
- collection_has_data
- add_collection_to_cache

Helper functions moved to module level:
- delete_prefix: Delete all keys with a given prefix
- extract_last_path_segment_str: Extract last segment from path

database.rs reduced from 1537 to 455 lines.
